Test for point being on line segment allows closeness
This is especially important for horizontal and vertical lines and one
looks for intersections.

+++ BezierLab/FLPathSegment.h
@@ -44,7 +44,7 @@
 @property (readonly) NSPoint midPoint;
 @property (strong) NSMutableArray *clippings;
 @property (assign) BOOL keep;
-@property (assign) int crossnum;
+@property (assign) NSUInteger crossnum;
 
 @end
